+++ changelog.md
@@ -26,3 +26,9 @@
 ## 1.2.3
 
 * PHPCBF
+
+## 1.3.0
+
+* Set `MathsHelper` to use standard rounding by default
+* Ensure proper rounding of whole numbers
+* Add unit tests for `MathsHelper`